requirements.txt
Flask==2.3.3
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-api-python-client==2.97.0
pytest==7.4.0
PyYAML==6.0.1
Flask-Cors==4.0.0
flake8==6.1.0
black==23.7.0
mypy==1.5.1